feat: implement user session checks and redirect in drawer and tabs layout to ensure access control based on authentication
